packages/core/image_upload/requirements.txt
pyjwt==2.7.0
